# src/web/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Clean Architecture Template",
    "version": "v1"
  },
  "paths": {
    "/api/v1/posts": {
      "post": {
        "description": "Create a new post",
        "tags": ["Posts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "published": {
                    "type": "boolean"
                  },
                  "title": {
                    "type": "string"
                  }
                },
                "required": ["published", "title"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created post",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": "1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{id}": {
      "delete": {
        "description": "Delete a post",
        "tags": ["Posts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the post to delete",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted post",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "The post does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Get a post",
        "tags": ["Posts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the post to retrieve",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved post",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": "1"
                    },
                    "published": {
                      "type": "boolean",
                      "example": true
                    },
                    "title": {
                      "type": "string",
                      "example": "Example post"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The post does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem"
          },
          "errors": {
            "type": "array",
            "description": "Exception details (e.g. validation result)",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "number",
            "description": "The HTTP status code"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type"
          },
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type"
          }
        },
        "required": ["status", "title", "type"]
      },
      "BadRequestResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ExceptionResponse"
          }
        ],
        "example": {
          "status": 400,
          "title": "The request was invalid",
          "type": "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        }
      },
      "NotFoundResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ExceptionResponse"
          }
        ],
        "example": {
          "status": 404,
          "title": "The specified resource was not found",
          "type": "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.4"
        }
      },
      "InternalErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ExceptionResponse"
          }
        ],
        "example": {
          "status": 500,
          "title": "An error occurred while processing your request",
          "type": "https://tools.ietf.org/html/rfc7231#section-6.6.1"
        }
      }
    }
  }
}